Test for equivalence of Motion-stepped time and Sequence time.
Constant stepping requires double-precision Time type.
Non-constant (while) is equivalent by definition (and by testing).

+++ src/choreograph/Source.hpp
@@ -34,9 +34,8 @@ namespace choreograph
 
 /// Define our Time type here so it's easier to change out if needed.
 /// Float loses precision pretty quickly, but is fast and doesn't take up much space.
-/// Sub in double if you find yourself needing extra precision for time.
-/// Also, since only Motions keep playheads, we only need enough precision for our longest Motion.
-using Time = float;
+/// Double works nicely when adding together lots of little bits.
+using Time = double;
 
 template<typename T>
 class Sequence;
